+++ changelog.md
@@ -1,3 +1,6 @@
+=================================
+- Added ContentEditor
+
 2.10.0
 =================================
 - Added EditableObject::delete()